+++ NEWS.md
@@ -1,5 +1,10 @@
 # SpectraVis 0.3
 
+## Changes in 0.3.1
+
+- Import `PSMatch::labelFragments()`.
+- Drop `msdata` for `MsDataHub`.
+
 ## Changes in 0.3.0
 
 - Add `plotlySpectraMirror` function.